Stage1씬에 바닥 구현, Intro, Ending 씬 추가

+++ WinAPI_BubbleBoggle/WinAPI_BubbleBoggle/BackGround.cpp
@@ -11,7 +11,7 @@ namespace MomDra
 		SelectGDI color{ hdc, PenType::Red };
 		
 		static const Vector2& resolution{ Core::GetInstance().GetResolution() };
-		static RECT rect{ 0, 0, resolution.X, resolution.Y };
+		static RECT rect{ 0, 0, static_cast<LONG>(resolution.X), static_cast<LONG>(resolution.Y) };
 
 		FillRect(hdc, &rect, HBRUSH(color.GetPen()));
 	}
